geneve: Fix crash in SerializeTo when lengths are wrong
We run the risk of panicing on buffer overflow if option lengths in the
packet are blindly trusted in serialization.

In line with the precedent set elsewhere (e.g. IPv4, TCP), compute the
necessary buffer size based on the actual options size.

Fill lengths fields from user-provided values, even if those are wrong.

Signed-off-by: Nick Zavaritsky <mejedi@gmail.com>