Fix DHCPv4 options encoding
For calculating the DHCPv4 length by the "Len" method there is always a
byte added for the end option. But the END option is only set when there
are any options at all. This results in a missing END option in case
there are no options present.

This commit moves the encoding of the END option so it is always set
unconditionally.