Fixes for Geneve option flags/length field (#17)
* Fix Geneve option flags/length field decoding

The option header length is 5 bits, not 4, so fix the bit shift for
decoding flags, and fix the mask for decoding the length. This also adds
a unit test for decoding a max length option (124 byte payload).

* Fix Geneve option flags/length field serialization

OR-assigning the flag bits into the byte offset for the Geneve option
header flags/length field produced corrupt packets because the buffer
returned by PrependBytes is not guaranteed to be zeroed out. Since this
is the first time that offset is touched, just assign instead.

---------

Co-authored-by: Simeon Miteff <simeon.miteff@corelight.com>