Fix DHCPv4 header field name hops (#21)
The DHCPv4 header field "HardwareOpts" is misleading. This fields
original name by RFC 2131, which just takes it from RFC 951, is "hops".
It has never changed meaning and still reflects number of relay agents
it passed, as described in RFC 951 section 8.

This commit changes the name of the field "HardwareOpts" to "RelayHops".