Fix RadioTap decoding panic (#45)
On rare occasions, it is possible that the length field in the RadioTap
layer is bigger than the actual data length which results in a panic
when trying to reference the payload. If it's bigger, simply truncate
the RadioTap length to the length of the actual data to allow processing
of what we can safely.