As stated in "RFC3261 - 20.14 Content-Length" the header must be used to determine message-size when transporting SIP over TCP.
The Content-Length header is utilized to determine the payload size
if it is present, otherwise payload size will be determined by the
size of the buffer being parsed.

The setTruncated flag will also be set during parsing if the buffer
being parsed does not contain enough data, to honor the Content-Length
header.