Using Content-Length to determine SIP message size (#51)
* As stated in "RFC3261 - 20.14 Content-Length" the header must be used
to determine message-size when transporting SIP over TCP.

The Content-Length header is utilized to determine the payload size
if it is present, otherwise payload size will be determined by the
size of the buffer being parsed.

The setTruncated flag will also be set during parsing if the buffer
being parsed does not contain enough data, to honor the Content-Length
header.

* Ensuring compatibility of .GetContentLength()

To ensure backwards compatibility, get .GetContentLength() method will no longer return the actual size of the payload, but ONLY reflect the content of the "Content-Length" header, and returning 0 if the header is not set.

Also updated tests to ensure the above.

* Adding one more test on Absent Content-Length header

This test will show that length of the message body is only limited if the Content-Length header is present.