Changing internal type of CSEQ and Content-Length
Prior to this int64 have been used to store the values, this requires
some, perhaps unsafe "truncation" of contentLength when mapping to the internal buffer for storing payload.

I have changed the internal type to better fit the usage of the values, and added guard checks when parsing.

To better test for parsing-errors I have refactored testing.