afpacket: provide access to the file descriptor of the AF_PACKET socket
This adds a FD() method to the afpacket.TPacket type to allow consumers
of the package to retrieve its internal file descriptor. This is
inspired by the similar (*File).Fd() method from the os package.

Some use cases require access to the file descriptor of the AF_PACKET
socket, e.g. to be able to reference the socket in other syscalls.